requirements.txt
astroid==3.1.0
attrs==23.2.0
aws-cdk-lib
aws-cdk.asset-awscli-v1
aws-cdk.asset-kubectl-v20
aws-cdk.asset-node-proxy-agent-v6
aws-cdk.assets
aws-cdk.aws-cloudwatch
aws-cdk.aws-ec2
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.aws-logs
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-ssm
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info
awscli
botocore==1.34.80
cattrs==23.2.3
colorama==0.4.4
constructs==10.3.0
dill==0.3.8
docutils==0.16
flake8==7.0.0
gitdb==4.0.11
GitPython==3.1.43
importlib_resources==6.4.0
isort==5.13.2
jmespath==1.0.1
jsii==1.97.0
lint==1.2.1
mccabe==0.7.0
platformdirs==4.2.0
publication==0.0.3
pyasn1==0.6.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.1.0
python-dateutil==2.9.0.post0
PyYAML
rsa
s3transfer
six
smmap
tomlkit
typeguard
typing_extensions
urllib3
